requirements.txt
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
